OSD: Check that we're supposed to be getting a PG before waitlisting requests.
This was broken in fa722de6708d3e92037df6289cc29ece12c8ea66. Fix it by
adding a new check_op_mapping() function and calling it when we don't
have the PG for a request.
Rework handle_misdirected_op slightly so it doesn't require us to have
a PG object for misdirected ops, and fix the documentation for
op_is_queueable() to not be a gigantic lie.

Signed-off-by: Greg Farnum <gregory.farnum@dreamhost.com>