librados: expose CephContext via C API
We can already create rados cluster handles with an existing CephContext,
but that is only useful if you are building something that has access to
ceph internals; the cct isn't exposed via the API itself.

Do so, for both teh cluster and pool handles.  Add cluster handle accessor
for the C++ API too.

Fixes: #1821
Signed-off-by: Sage Weil <sage@newdream.net>